Add fill_value option in interpolateion
`interpolate_contour` may return nan when failed to interpolate grid
array in the fit. This fix adds `fill_value` option to make it
accessible to fill in the failed points with a given value. Also, when
extrapolation in the fit occures with the default `fill_value`, a
warning message to suggest to use `fill_value` is displayed.